pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-team"
version = "0.1.0"
description = "Convergence-driven multi-agent orchestration system using Claude Agent SDK"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Omar Khaled"},
]

dependencies = [
    "claude-agent-sdk>=0.1.26",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
agent-team = "agent_team.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring real API keys",
    "integration: cross-module integration tests",
]

[tool.setuptools.packages.find]
where = ["src"]
